moved initial sidebar visibility logic frop app component to sidebar service
