+++ src/app/app.component.ts
@@ -1,4 +1,4 @@
-﻿﻿import { Component, OnInit, Inject, NgZone } from '@angular/core';
+﻿﻿import { Component, OnInit, Inject } from '@angular/core';
 import { ROUTER_DIRECTIVES, RouteConfig, Router } from '@angular/router-deprecated';
 
 import { TranslateService, TranslatePipe } from "ng2-translate/ng2-translate";
@@ -151,8 +151,6 @@ export class AppComponent implements OnInit {
      *      Router is a singleton service provided by Angular2.
      * @param sidebarService
      *      SidebarService is a singleton service that provides access to the content of the sidebar
-     * @param viewportService
-     *      A singleton service that classifies the viewport's width
      * @param sidebarHelper
      *      SidebarHelper is a helper-class to inject the sidebar sections when the user visits this component
      */
@@ -160,7 +158,6 @@ export class AppComponent implements OnInit {
                 private translate:TranslateService,
                 private router:Router,
                 private sidebarService: SidebarService,
-                private viewportService: ViewportService,
                 @Inject(AppSidebarHelper)  private sidebarHelper:AppSidebarHelper) {
         translate.setDefaultLang('en');
         translate.use('en');
@@ -173,6 +170,7 @@ export class AppComponent implements OnInit {
      */
     private initSidebar(): void {
         this.isSidebarAnimating = false;
+        this.isSidebarVisible = this.sidebarService.isSidebarVisible.getValue();
 
         //Subscribe to the visibility observable of the SidebarService to
         //detect when the sidebar should open and close
@@ -195,28 +193,7 @@ export class AppComponent implements OnInit {
             }, 350);
         });
 
-        //if we have info about the viewport
-        //use it to determine the initial state of the sidebar 
-        if (this.viewportService.isSupported) {
-            let isMd = this.viewportService.isMd.getValue();
-            let isLg = this.viewportService.isLg.getValue();
-            this.sidebarService.setSidebarVisibility(isLg || isMd);
-        }
-        else {
-            //otherwise, default to closed.
-            this.sidebarService.setSidebarVisibility(false);
-        }
-
         this.sidebarHelper.populateSidebar();
-
-        this.router.subscribe(() => {
-            // if the route changes on an XS screen (where the sidebar is fullscreen),
-            // and the sidebar is open, close it.
-            if (this.viewportService.isXs.getValue() &&
-                this.sidebarService.isSidebarVisible.getValue()) {
-                this.sidebarService.setSidebarVisibility(false);
-            }
-        });
     }
 
     /**
